Rename the flare folder flare_events
