+++ backend/pe/src/pe_source/flare_events/__init__.py
@@ -6,7 +6,7 @@ __all__ = ["run_flare_events"]
 def __getattr__(name):
     if name == "run_flare_events":
         # Third-Party Libraries
-        from pe_source.flare.flare_events_script import run_flare_events
+        from pe_source.flare_events.flare_events_script import run_flare_events
 
         return run_flare_events
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
